model_council_app/requirements.txt
flask
ollama
numpy
pypdf2
python-docx
requests
beautifulsoup4
